requirements.txt
accelerate==0.27.2
deepspeed==0.10.0
hf_transfer==0.1.8
huggingface_hub==0.24.6
jsonlines==4.0.0
mmcv==1.7.0
numpy==2.2.0
openai==1.57.2
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
protobuf==3.20.0
rich==13.9.4
scikit_learn==1.2.2
setuptools==69.5.1
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.21.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.4
transformers==4.46.3
xformers==0.0.28.post3
